+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [1.0.18]
+
+* Do not auto focus gpt model input field when opening menu drawer
+* Fixed error is shown in system user's message when failed to send message or OpenAI API returns error
+* Added auto-suggestion for message input field
+
 ## [1.0.17]
 
 * Removed save conversation to file icon in AppBar